nidas::core::SampleScanner:     Changed code which estimates next sample size.     Changed buffer size for MessageStreamScanners from 1024 to 2048.     Differential GPS has samples of size 1314. Increasing the buffer     size will make this a bit more efficient.     In MessageStreamSampleScanner, When a sample must be enlarged,     get a new one from the pool, rather than reallocating the data portion     of the current sample. nidas::core::SamplePool:     Handle possibility of sample from small pool being returned     to large pool.  Added methods to query pool sizes. nidas::core::StatusHandler,StatusListener,StatusThread:     Report sample pool sizes in DSMEngine status in a new <samplepool>     element.  Ran indent on these with indent of 4. nidas::core::DSMSensor:     sample sizes are unsigned int, not size_t.

+++ src/nidas/core/DSMSensor.h
@@ -703,13 +703,13 @@ public:
         if (_scanner) _scanner->calcStatistics(periodUsec);
     }
 
-    size_t getMaxSampleLength() const
+    unsigned int getMaxSampleLength() const
     {
 	if (_scanner) return _scanner->getMaxSampleLength();
 	return 0;
     }
 
-    size_t getMinSampleLength() const
+    unsigned int getMinSampleLength() const
     {
         if (_scanner) return _scanner->getMinSampleLength();
 	return 0;
